[FIX] stock_location_orderpoint_change_priority: Correctly update replenishment move priority
Previously, when a replenishment move's priority was updated, the `stock.move` record's priority did not change. This was because `stock.move.priority` is a computed field that depends on the associated `stock.picking`'s priority.

This commit fixes the issue by updating the picking's priority as well.

This also includes:
- Passing the `product` argument to the `super().run_replenishment` method.
- Updating the module's manifest description.